Change build system to make
The ant scripts included in the android SDK were removed recently. The
alternative is Gradle.
Gradle doesn't work well in this app because it's not possible to use
Nix to pull dependencies. Gradle will try to patch the SDK.
Also, it's very slow.

It turns out the required build script is quite simple.