custom labels in xml layout
allow for defining optional custom labels inside the <key> element of xml layouts using key0_label and key1_label...key8_label or their synonyms nw_label...s_label.

currently, this is mostly useful for keys with multiple characters on key1...key8, since the automatic label might be too long on key0 or not unique on key1...key8. the custom labels might also be useful as a workaround on key1...key8, where automatic labels are cut off after 3 characters instead of 3 visible symbols, which can lead to strange results with emojis or other unicode combinations.
in the future, custom labels might be neccessary, when key combinations/sequences or a "goto layout" button should be implemented.

the labels do not interfere with the functionality of the keys. for example, the custom label is not changed by modifiers, even if the underlying key still is.
on key0 custom labels longer than 1 character are drawn using the reduced font size in the same way as automatic labels.
on key1...key8 custom labels are not truncated after 3 characters. you can define the label in the way it fits your layout.